src/requirements.txt
pytest
fastapi
PyJWT
pydantic
pydantic[email]
bcrypt
pymongo
python-dotenv
pytest-mock
mangum
bcrypt